refactor(oh7): fold parser into core::parser (#252)
