+++ src/core/mod.rs
@@ -10,6 +10,7 @@ pub mod model_updater;
 #[allow(dead_code)]
 pub mod orchestration;
 pub mod pack_validation;
+pub mod parser;
 pub mod project;
 pub mod project_registry;
 pub mod prompt;
