+++ src/core/parser/metadata.rs
