+++ src/core/prompt.rs
@@ -2,7 +2,7 @@ use std::path::{Path, PathBuf};
 
 use serde::Deserialize;
 
-use crate::parser::frontmatter::extract_frontmatter;
+use crate::core::parser::frontmatter::extract_frontmatter;
 
 // ---------------------------------------------------------------------------
 // Data model
